organized pages directory and added loaders
